Decoupling other functions from views
  new file:  src/djangosaml2_spid/spid_metadata.py
  new file:  src/djangosaml2_spid/spid_request.py
  modified:  src/djangosaml2_spid/views.py
