+++ src/djangosaml2_spid/views.py
@@ -1,4 +1,6 @@
 from .spid_anomalies import SpidAnomaly
+from .spid_metadata import spid_sp_metadata
+from .spid_request import spid_sp_authn_request
 from .utils import repr_saml
 from django.conf import settings
 from django.contrib import auth
@@ -11,7 +13,6 @@ from django.http import (
     HttpResponseRedirect
 )
 from django.shortcuts import render
-from django.urls import reverse
 from djangosaml2.cache import IdentityCache, OutstandingQueriesCache
 from djangosaml2.cache import StateCache
 from djangosaml2.conf import get_config
@@ -22,11 +23,8 @@ from djangosaml2.utils import (
     validate_referral_url
 )
 from saml2 import BINDING_HTTP_REDIRECT, BINDING_HTTP_POST
-from saml2.authn_context import requested_authn_context
 from saml2.mdstore import UnknownSystemEntity
-from saml2.metadata import entity_descriptor, sign_entity_descriptor
 from saml2.s_utils import UnsupportedBinding
-from saml2.sigver import security_context
 import djangosaml2.views as djangosaml2_views
 import logging
 import saml2
@@ -39,7 +37,7 @@ logger = logging.getLogger('djangosaml2')
 
 
 def index(request):
-    "Barebone 'diagnostics' view, print user attributes if logged in + login/logout links."
+    """Barebone 'diagnostics' view, print user attributes if logged in + login/logout links."""
 
     if request.user.is_authenticated:
         out = f"LOGGED IN: <a href={settings.LOGOUT_URL}>LOGOUT</a><br>"
@@ -53,76 +51,8 @@ def index(request):
         return HttpResponse(f"LOGGED OUT: <a href={settings.LOGIN_URL}>LOGIN</a>")
 
 
-def spid_sp_authn_request(conf, selected_idp, next_url=''):
-    client = Saml2Client(conf)
-
-    logger.debug(f'Redirecting user to the IdP via {SPID_DEFAULT_BINDING} binding.')
-
-    # use the html provided by pysaml2 if no template was specified or it didn't exist
-    # SPID want the fqdn of the IDP, not the SSO endpoint
-    location_fixed = selected_idp
-    location = client.sso_location(selected_idp, SPID_DEFAULT_BINDING)
-
-    authn_req = saml2.samlp.AuthnRequest()
-    authn_req.destination = location_fixed
-    # spid-testenv2 preleva l'attribute consumer service dalla authnRequest (anche se questo sta già nei metadati...)
-    authn_req.attribute_consuming_service_index = "0"
-
-    # issuer
-    issuer = saml2.saml.Issuer()
-    issuer.name_qualifier = client.config.entityid
-    issuer.text = client.config.entityid
-    issuer.format = "urn:oasis:names:tc:SAML:2.0:nameid-format:entity"
-    authn_req.issuer = issuer
-
-    # message id
-    authn_req.id = saml2.s_utils.sid()
-    authn_req.version = saml2.VERSION  # "2.0"
-    authn_req.issue_instant = saml2.time_util.instant()
-
-    name_id_policy = saml2.samlp.NameIDPolicy()
-    name_id_policy.format = settings.SPID_NAMEID_FORMAT
-    authn_req.name_id_policy = name_id_policy
-
-    authn_context = requested_authn_context(class_ref=settings.SPID_AUTH_CONTEXT)
-    authn_req.requested_authn_context = authn_context
-
-    # if SPID authentication level is > 1 then forceauthn must be True
-    authn_req.force_authn = 'true'
-
-    authn_req.protocol_binding = SPID_DEFAULT_BINDING
-
-    assertion_consumer_service_url = client.config._sp_endpoints['assertion_consumer_service'][0][0]
-    authn_req.assertion_consumer_service_url = assertion_consumer_service_url
-
-    authn_req_signed = client.sign(
-        authn_req,
-        sign_prepare=False,
-        sign_alg=settings.SPID_SIG_ALG,
-        digest_alg=settings.SPID_DIG_ALG,
-    )
-
-    logger.debug(f'AuthRequest to {selected_idp}: {authn_req_signed}')
-
-    relay_state = next_url or reverse('djangosaml2:saml2_echo_attributes')
-    http_info = client.apply_binding(
-        SPID_DEFAULT_BINDING,
-        authn_req_signed,
-        location,
-        sign=True,
-        sigalg=settings.SPID_SIG_ALG,
-        relay_state=relay_state
-    )
-
-    return dict(
-        http_response=http_info,
-        authn_request=authn_req_signed,
-        relay_state=relay_state,
-        session_id=authn_req.id
-    )
-
-
-def spid_login(request, config_loader_path=None, wayf_template='wayf.html', authorization_error_template='djangosaml2/auth_error.html'):
+def spid_login(request, config_loader_path=None, wayf_template='wayf.html',
+               authorization_error_template='djangosaml2/auth_error.html'):
     """SAML Authorization Request initiator
 
     This view initiates the SAML2 Authorization handshake
@@ -312,141 +242,6 @@ def spid_logout(request, config_loader_path=None, **kwargs):
     )
     state.sync()
     return HttpResponse(http_info['data'])
-
-
-def spid_sp_metadata(conf):
-    metadata = entity_descriptor(conf)
-
-    # this will renumber acs starting from 0 and set index=0 as is_default
-    cnt = 0
-    for attribute_consuming_service in metadata.spsso_descriptor.attribute_consuming_service:
-        attribute_consuming_service.index = str(cnt)
-        cnt += 1
-
-    cnt = 0
-    for assertion_consumer_service in metadata.spsso_descriptor.assertion_consumer_service:
-        assertion_consumer_service.is_default = 'true' if not cnt else ''
-        assertion_consumer_service.index = str(cnt)
-        cnt += 1
-
-    # nameformat patch
-    for reqattr in metadata.spsso_descriptor.attribute_consuming_service[0].requested_attribute:
-        reqattr.name_format = None  # "urn:oasis:names:tc:SAML:2.0:attrname-format:basic"
-        reqattr.friendly_name = None
-
-    metadata.extensions = None
-
-    # attribute consuming service service name patch
-    service_name = metadata.spsso_descriptor.attribute_consuming_service[0].service_name[0]
-    service_name.lang = 'it'
-    service_name.text = conf._sp_name
-
-    avviso_29_v3(metadata)
-
-    # metadata signature
-    secc = security_context(conf)
-    sign_dig_algs = dict(
-        sign_alg=conf._sp_signing_algorithm,
-        digest_alg=conf._sp_digest_algorithm
-    )
-    eid, xmldoc = sign_entity_descriptor(metadata, None, secc, **sign_dig_algs)
-    return xmldoc
-
-
-def avviso_29_v3(metadata):
-    "https://www.agid.gov.it/sites/default/files/repository_files/spid-avviso-n29v3-specifiche_sp_pubblici_e_privati_0.pdf"
-
-    saml2.md.SamlBase.register_prefix(settings.SPID_PREFIXES)
-
-    contact_map = settings.SPID_CONTACTS
-    metadata.contact_person = []
-    for contact in contact_map:
-        spid_contact = saml2.md.ContactPerson()
-        spid_contact.contact_type = contact['contact_type']
-        contact_kwargs = {
-            'email_address': [contact['email_address']],
-            'telephone_number': [contact['telephone_number']]
-        }
-        if contact['contact_type'] == 'other':
-            spid_contact.loadd(contact_kwargs)
-            contact_kwargs['contact_type'] = contact['contact_type']
-            spid_extensions = saml2.ExtensionElement(
-                'Extensions',
-                namespace='urn:oasis:names:tc:SAML:2.0:metadata'
-            )
-            for k, v in contact.items():
-                if k in contact_kwargs:
-                    continue
-                ext = saml2.ExtensionElement(
-                    k,
-                    namespace=settings.SPID_PREFIXES['spid'],
-                    text=v
-                )
-                spid_extensions.children.append(ext)
-
-        elif contact['contact_type'] == 'billing':
-            contact_kwargs['company'] = contact['company']
-            spid_contact.loadd(contact_kwargs)
-            spid_extensions = saml2.ExtensionElement(
-                'Extensions',
-                namespace='urn:oasis:names:tc:SAML:2.0:metadata'
-            )
-
-            elements = {}
-            for k, v in contact.items():
-                if k in contact_kwargs:
-                    continue
-                ext = saml2.ExtensionElement(
-                    k,
-                    namespace=settings.SPID_PREFIXES['fpa'],
-                    text=v
-                )
-                elements[k] = ext
-
-            # DatiAnagrafici
-            IdFiscaleIVA = saml2.ExtensionElement(
-                'IdFiscaleIVA',
-                namespace=settings.SPID_PREFIXES['fpa'],
-            )
-            Anagrafica = saml2.ExtensionElement(
-                'Anagrafica',
-                namespace=settings.SPID_PREFIXES['fpa'],
-            )
-            Anagrafica.children.append(elements['Denominazione'])
-
-            IdFiscaleIVA.children.append(elements['IdPaese'])
-            IdFiscaleIVA.children.append(elements['IdCodice'])
-            DatiAnagrafici = saml2.ExtensionElement(
-                'DatiAnagrafici',
-                namespace=settings.SPID_PREFIXES['fpa'],
-            )
-            if elements.get('CodiceFiscale'):
-                DatiAnagrafici.children.append(elements['CodiceFiscale'])
-            DatiAnagrafici.children.append(IdFiscaleIVA)
-            DatiAnagrafici.children.append(Anagrafica)
-            CessionarioCommittente = saml2.ExtensionElement(
-                'CessionarioCommittente',
-                namespace=settings.SPID_PREFIXES['fpa'],
-            )
-            CessionarioCommittente.children.append(DatiAnagrafici)
-
-            # Sede
-            Sede = saml2.ExtensionElement(
-                'Sede',
-                namespace=settings.SPID_PREFIXES['fpa'],
-            )
-            Sede.children.append(elements['Indirizzo'])
-            Sede.children.append(elements['NumeroCivico'])
-            Sede.children.append(elements['CAP'])
-            Sede.children.append(elements['Comune'])
-            Sede.children.append(elements['Provincia'])
-            Sede.children.append(elements['Nazione'])
-            CessionarioCommittente.children.append(Sede)
-
-            spid_extensions.children.append(CessionarioCommittente)
-
-        spid_contact.extensions = spid_extensions
-        metadata.contact_person.append(spid_contact)
 
 
 def metadata_spid(request, config_loader_path=None, valid_for=None):
